scripts/requirements.txt
greenlet==3.0.3
numpy==2.1.1
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-environ==0.4.54
pytz==2024.1
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
